+++ frontend/src/App.jsx
@@ -3,43 +3,42 @@ import {
   Routes,
   Route,
   Navigate,
-} from "react-router-dom";
+} from "react-router-dom"
 
 //----------------------------------GUEST---------------------------------------------
-import AuthLayout from "./layouts/AuthLayout";
-import LandingPage from "./components/CommonComponents/LandingPage/LandingPage.jsx";
-import Login from "./components/CommonComponents/Login/Login.jsx";
-import Register from "./components/CommonComponents/Register/Register.jsx";
-import ForgotPassword from "./components/CommonComponents/ForgotPassword/ForgotPassword.jsx";
-import NewPassword from "./components/CommonComponents/NewPassword/NewPassword.jsx";
-import ConfirmAccount from "./components/CommonComponents/ConfirmAccount/ConfirmAccount.jsx";
-import UpdatePassword from "./components/CommonComponents/UpdatePassword/UpdatePassword.jsx";
-import Details from "./components/CommonComponents/Details/Details.jsx";
-import About from "./components/CommonComponents/About/About";
-import NotFound from "./components/CommonComponents/NotFound/NotFound.jsx";
-import Home from "./components/CommonComponents/Home/Home";
-import Homeout from "./components/CommonComponents/Homeout/Homeout.jsx";
-import UserCard from "./components/CommonComponents/UserCard/UserCard.jsx";
+import AuthLayout from "./layouts/AuthLayout"
+import LandingPage from "./components/CommonComponents/LandingPage/LandingPage.jsx"
+import Login from "./components/CommonComponents/Login/Login.jsx"
+import Register from "./components/CommonComponents/Register/Register.jsx"
+import ForgotPassword from "./components/CommonComponents/ForgotPassword/ForgotPassword.jsx"
+import NewPassword from "./components/CommonComponents/NewPassword/NewPassword.jsx"
+import ConfirmAccount from "./components/CommonComponents/ConfirmAccount/ConfirmAccount.jsx"
+import UpdatePassword from "./components/CommonComponents/UpdatePassword/UpdatePassword.jsx"
+import Details from "./components/CommonComponents/Details/Details.jsx"
+import About from "./components/CommonComponents/About/About"
+import NotFound from "./components/CommonComponents/NotFound/NotFound.jsx"
+import Home from "./components/CommonComponents/Home/Home"
+import Homeout from "./components/CommonComponents/Homeout/Homeout.jsx"
+import UserCard from "./components/CommonComponents/UserCard/UserCard.jsx"
 //----------------------------------USER---------------------------------------------
-import VerificationUser from "./components/CommonComponents/VerificationUser/VerificationUser.jsx";
-import RutaProtegida from "./layouts/RutaProtegida";
-import Forms from "./components/UserRegisteredComponents/Form/Forms.jsx";
-import Settings from "./components/UserRegisteredComponents/Settings/Settings.jsx";
-import UpdateBook from "./components/UserRegisteredComponents/UpdateBook/UpdateBook.jsx";
-import CheckOut from "./components/UserRegisteredComponents/CheckOut/CheckOut";
-import OrdersHistory from "./components/UserRegisteredComponents/OrdersHistory/OrdersHistory";
-import OrderDetail from "./components/UserRegisteredComponents/OrderDetail/OrderDetail";
+import VerificationUser from "./components/CommonComponents/VerificationUser/VerificationUser.jsx"
+import RutaProtegida from "./layouts/RutaProtegida"
+import Forms from "./components/UserRegisteredComponents/Form/Forms.jsx"
+import Settings from "./components/UserRegisteredComponents/Settings/Settings.jsx"
+import CheckOut from "./components/UserRegisteredComponents/CheckOut/CheckOut"
+import OrdersHistory from "./components/UserRegisteredComponents/OrdersHistory/OrdersHistory"
+import OrderDetail from "./components/UserRegisteredComponents/OrderDetail/OrderDetail"
 import BooksCreated from "./components/UserRegisteredComponents/BooksCreated/BooksCreated"
-import ProfileUsers from "./components/CommonComponents/ProfileUsers/ProfileUsers";
+import ProfileUsers from "./components/CommonComponents/ProfileUsers/ProfileUsers"
 import Review from './components/UserRegisteredComponents/Review/Review'
 import HomeChat from './components/CommonComponents/Chat/HomeChat' 
 //----------------------------------ADMIN---------------------------------------------
-import AllUsers from "./components/AdminComponents/AllUsers/AllUsers";
-import AdminHome from "./components/AdminComponents/AdminHome/AdminHome";
-import AllOrders from "./components/AdminComponents/AllOrders/AllOrders";
-import ConfirmacionPago from "./components/UserRegisteredComponents/CheckOut/ConfirmacionPago/ConfirmacionPago";
-import Questions from "./components/UserRegisteredComponents/Questions/Questions.jsx";
-import Profile from "./components/UserRegisteredComponents/Profile/Profile";
+import AllUsers from "./components/AdminComponents/AllUsers/AllUsers"
+import AdminHome from "./components/AdminComponents/AdminHome/AdminHome"
+import AllOrders from "./components/AdminComponents/AllOrders/AllOrders"
+import ConfirmacionPago from "./components/UserRegisteredComponents/CheckOut/ConfirmacionPago/ConfirmacionPago"
+import Questions from "./components/UserRegisteredComponents/Questions/Questions.jsx"
+import Profile from "./components/UserRegisteredComponents/Profile/Profile"
 import Answers from "./components/UserRegisteredComponents/Answers/Answers"
 
 
@@ -70,7 +69,7 @@ export default function App() {
           <Route path="/questions/:id" element={<Questions />}/>
           <Route path="/answers/:id" element={<Answers />}/>
           <Route path="/profile" element={<Profile />}/>
-          <Route path="/details/update/:id" element={<UpdateBook />} />
+          <Route path="/details/update/:id" element={<Forms />} />
           <Route path="/details/:id" element={<Details />} />
           <Route path="/user/setting" element={<Settings />} />
           <Route path="/checkout" element={<CheckOut />} />
@@ -84,9 +83,8 @@ export default function App() {
           <Route path="/chat" element={<HomeChat />} />
           </Route>
 
-
         <Route path="*" element={<NotFound />} />
       </Routes>
     </Router>
-  );
+  )
 }
