+++ frontend/src/components/CommonComponents/Details/DetailsField/DetailsField.jsx
@@ -1,0 +1,19 @@
+const DetailsField = (
+  {
+    constant,
+    clase,
+    title
+  }) => {
+  return (
+    <div>
+      {
+        constant ?
+          (<>
+            <h5 className={clase}>{title}</h5> {constant}
+          </>)
+          : null
+      }
+    </div>)
+}
+
+export default DetailsField
