+++ frontend/src/components/UserRegisteredComponents/Form/EditCard/EditCard.jsx
@@ -1,16 +1,18 @@
-import { useSelector } from "react-redux"
+import { useEffect } from "react"
+import { useDispatch, useSelector } from "react-redux"
 import { detailsBook } from "../../../../redux/actions/detailsBooks"
 
-const { nombre, autor, idioma, editorial, edicion, tapa, año_de_pub, cant_pags, descripcion, price, file, colection, ilustrado, category } = useSelector(state => state.detail)
-
 const EditCard = ({ id, addMode }) => {
+  const dispatch = useDispatch()
+  const { nombre, autor, idioma, editorial, edicion, tapa, año_de_pub, cant_pags, descripcion, price, file, colection, ilustrado, category } = useSelector(state => state.detail)
+
   useEffect(() => {
     if (!addMode) { dispatch(detailsBook(id)) }
   }, [])
 
   return (
     <div>
-      <h1>{addMode ? 'Anunciar Producto' : `Editar ${nombre}`}</h1>
+      <h1>{addMode ? 'Anunciar Producto' : `Editar: ${nombre}`}</h1>
       {
         addMode
           ? null
