+++ frontend/src/components/UserRegisteredComponents/Form/Forms.jsx
@@ -7,16 +7,16 @@ import { getCategories } from '../../../redux/actions/actionCategories.js'
 import { cleanData, getBooks, putBookBody } from '../../../redux/actions/actionBooks'
 import { detailsBook } from '../../../redux/actions/detailsBooks'
 import { subirFotos } from '../../../redux/actions/actionSubirFotos'
-import { currentYear } from '../../../utils/helperFunctions'
 import PreviewImage from './ImgPreview/ImgPreview'
 import NavBar from '../../CommonComponents/NavBar/NavBar'
 import s from '../Form/Form.module.css'
+import EditCard from './EditCard/EditCard'
+import { formValidators } from '../../../utils/helperFunctions.js'
 
 const Forms = () => {
   const [dispatch, navigate] = [useDispatch(), useNavigate()]
   const categories = useSelector(state => state.categories)
   const imgPreview = useSelector(state => state.tempState)
-  const { nombre, autor, idioma, editorial, edicion, tapa, año_de_pub, cant_pags, descripcion, price, file, colection, ilustrado, category } = useSelector(state => state.detail)
   const { id } = useParams()
   const isAddMode = !id
 
@@ -37,28 +37,9 @@ const Forms = () => {
       <div>
         <NavBar />
 
-        <h1>{isAddMode ? 'Anunciar Producto' : `Editar ${nombre}`}</h1>
-        {!isAddMode
-          ? (
-            <p>{nombre}</p>,
-            <p>{autor}</p>,
-            <p>{idioma}</p>,
-            <p>{editorial}</p>,
-            <p>{edicion}</p>,
-            <p>{tapa}</p>,
-            <p>{año_de_pub}</p>,
-            <p>{cant_pags}</p>,
-            <p>{descripcion}</p>,
-            <p>{price}</p>,
-            <p>{file}</p>,
-            <p>{colection}</p>,
-            <p>{ilustrado}</p>,
-            <p>{category}</p>
-          )
-          : null
-        }
+        <EditCard id={id} addMode={isAddMode} />
 
-        < Formik
+        <Formik
           initialValues={{
             nombre: '',
             autor: '',
@@ -76,84 +57,18 @@ const Forms = () => {
             category: []
           }}
 
-          validate={(values) => {
-            let errors = {}
-
-            if (!values.nombre) {
-              errors.nombre = 'Campo requerido'
-            } else if (!/^\S.*$/.test(values.nombre)) {
-              errors.nombre = 'El primer caracter no puede ser un espacio'
-            } else if (!/^(\d|[a-z]|[\u00f1\u00d1]|[,.:¡!¿?']|[À-ÿ]|\s){1,40}$/i.test(values.nombre)) {
-              errors.nombre = 'Ingrese un nombre válido de hasta 40 caracteres'
-            }
-
-            if (!values.autor) {
-              errors.autor = 'Campo requerido'
-            } else if (!/^\S.*$/.test(values.autor)) {
-              errors.autor = 'El primer caracter no puede ser un espacio'
-            } else if (!/^(|[a-z]|[()']|[À-ÿ]|[\u00f1\u00d1]|\s){1,40}$/i.test(values.autor)) {
-              errors.autor = 'Ingrese un autor válido de hasta 40 caracteres'
-            }
-
-            if (!values.idioma) {
-              errors.idioma = 'Campo requerido'
-            } else if (!/^\S.*$/.test(values.idioma)) {
-              errors.idioma = 'El primer caracter no puede ser un espacio'
-            } else if (!/^([a-z]|[\u00f1\u00d1]|\s){1,20}$/i.test(values.idioma)) {
-              errors.idioma = 'Ingrese un idioma válido de hasta 40 caracteres'
-            }
-
-            if (/^\s(.)*$/.test(values.editorial)) {
-              errors.editorial = 'El primer caracter no puede ser un espacio'
-            } else if (!/^(\d|[a-z]|[\u00f1\u00d1]|[,.:¡!¿?']|[À-ÿ]|\s){0,40}$/i.test(values.editorial)) {
-              errors.editorial = 'Ingrese un nombre válido de hasta 40 caracteres'
-            }
-
-            if (/(\D)/.test(values.edicion) || values.edicion < 1 && values.edicion.toString().length > 0) {
-              errors.edicion = 'Ingrese un Nº de edición mayor a 0'
-            }
-
-            if (!/^([a-z]|\s){0,15}$/i.test(values.tapa)) {
-              errors.tapa = 'Ingrese un tipo de tapa'
-            }
-
-            if (values.año_de_pub && (!/^[0-9]{0,4}$/.test(values.año_de_pub) || values.año_de_pub > currentYear())) {
-              errors.año_de_pub = 'Ingrese un año válido en formato AAAA'
-            }
-
-            if (/(\D|^0|[-])/.test(values.cant_pags)) { // NO tira error si solo se le pasa "-"
-              errors.cant_pags = 'Ingrese un número de págs. válido'
-            }
-
-            if (values.descripcion.length < 6) {
-              errors.descripcion = 'La descripción debe contar con al menos 6 caracteres'
-            } else if (values.descripcion.length > 1500) {
-              errors.descripcion = 'La descripción debe contar con un máximo de 1500 caracteres'
-            }
-
-            if (/(\D)/.test(values.price)) {
-              errors.price = 'Ingrese el precio en centavos de USD'
-            } else if (!values.price || values.price < 50) {
-              errors.price = 'Ingrese un precio válido mayor a 50 centavos'
-            }
-
-            if (values.category.length < 1) {
-              errors.category = 'Elija al menos 1 categoría'
-            }
-
-            return errors
-          }}
+          validate={values => formValidators(values)}
 
           onSubmit={(values, { resetForm }) => {
             values.image = imgPreview
             delete values.file
-            
+
             if (isAddMode) { dispatch(postCreate(values)) }
             else {
-              values.id = id    
+              values._id = id
               dispatch(putBookBody(values))
-            }          
-            console.log("🚀 ~ file: Forms.jsx ~ values", values)
+            }
+            
             dispatch(cleanData)
             resetForm()
             swal({
@@ -167,7 +82,7 @@ const Forms = () => {
           }}
         >
 
-          {({ errors, handleSubmit, values, setFieldValue }) => (
+          {({ errors, values, handleSubmit, setFieldValue }) => (
             <Form onSubmit={handleSubmit} className={s.formik} >
               <div className={s.form}>
 
@@ -354,6 +269,7 @@ const Forms = () => {
                   type="submit"
                   disabled={errors.nombre || errors.autor || errors.idioma || errors.price || errors.category || errors.descripcion}
                 >ENVIAR</button>
+
               </div>
             </Form>
           )}
